A dockerfile for gapseq
Hi,

I have been facing issues with installing gapseq on different platforms but was able to install it on my Ubuntu. I decided to make a dockerfile and share it so, hopefully, saves time for other folks who face the same issues. I tested the container and it passes all the tests. It migh be also useful to build a pipeline with nextflow to allow large-scale GEM reconstruction in parallel on HPC systems.

Best,
Parsa